fix bookings page build error
